refactor: render AiTestEditor component in place of inline panel
